% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expand.MEANS.R
\name{expand.MEANS}
\alias{expand.MEANS}
\title{expand.MEANS}
\usage{
expand.MEANS(age.totals, age.labels, k = 30, year = "", weekly = TRUE,
  wts = NULL, tidy = TRUE, ...)
}
\arguments{
\item{age.totals}{the data variable of interest. It is the MEAN SPEND for each age-group}

\item{age.labels}{age.group categories}

\item{k}{the degrees of freedom in the generalized additive model}

\item{year}{a placeholder}

\item{weekly}{(logical)}

\item{wts}{the weights for the generalzied additive model}

\item{tidy}{(logical)}
}
\description{
expand.MEANS
}

