% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expand.SINGLE.R
\name{expand.SINGLE}
\alias{expand.SINGLE}
\title{expand.SINGLE}
\usage{
expand.SINGLE(age.totals, age.labels, k = 30, year = "", wts = NULL,
  tidy = TRUE, already.split = FALSE, ...)
}
\arguments{
\item{age.totals}{the data variable of interest, a total for each age bucket}

\item{age.labels}{the age group categories}

\item{k}{the degrees of freedom in the var}

\item{wts}{the weights for the generalized additive model}

\item{already.split}{(logical) is the data already in single year-of-age format}
}
\description{
expand.SINGLE
}

