% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/weighted.contribution.R
\name{weighted.contribution}
\alias{weighted.contribution}
\title{weighted contribution}
\usage{
weighted.contribution(vec, wts)
}
\arguments{
\item{vec}{}

\item{wts}{}
}
\description{
weighted contribution
}

